- refactor to extract koin setup for previews in a generic PreviewEnvironment composable to use for other complex UI previews
